Refactor sales_models into package; preserve import path
- Replace src/facts/sales/sales_models.py with a sales_models/ package.
- Split logic into focused modules: activity_model, customer_lifecycle, pricing_pipeline, quantity_model.
- Re-export public API from sales_models/__init__.py so downstream imports remain unchanged.
