+++ config.yaml
@@ -35,7 +35,7 @@ sales:
   chunk_size: 1000000
   max_lines_per_order: 10
   
-  file_format: "csv"          # csv | parquet | deltaparquet
+  file_format: "parquet"          # csv | parquet | deltaparquet
   sales_output: "sales"           # sales | sales_order | both
   write_delta: true
   delta_output_folder: "./data/fact_out/delta"
